backend/requirements.txt
django>=4.2
djangorestframework
django-cors-headers
PyYAML
channels
channels_redis
daphne
djangorestframework-simplejwt